default theme is changed to dark +in category.html implimanted sweetalert2 for confirmation

+++ admin/js/category.js
@@ -3,10 +3,11 @@ let product = JSON.parse(localStorage.getItem('products')) || "{}";
 let arr = []
 document.getElementById('addDescription').value = "";
 let data = JSON.parse(jsonString);
-const filter = document.getElementById('filter')
+const filter = document.getElementById('filter');
+upadateData();
+resetArr();
 displayElements(data);
 resetSortIcons();
-resetArr();
 function resetArr() {
     arr = [];
     for (let i in data) {
@@ -32,8 +33,14 @@ function removeEventListenersByClassName(className) {
     });
 }
 
-function displayElements() {
-    data = JSON.parse(localStorage.getItem('category') || "{}");
+function upadateData() {
+    jsonString = localStorage.getItem('category') || "{}";
+    product = JSON.parse(localStorage.getItem('products')) || "{}";
+    data = JSON.parse(jsonString);
+    resetArr();
+}
+
+function displayElements(data) {
     let tableBody = document.getElementById('tableBody');
     tableBody.innerHTML = "";
     for (let i in data) {
@@ -41,23 +48,25 @@ function displayElements() {
         let str = (element['description'].length > 50) ? element['description'].substring(0, 70) + "..." : element['description'];
         tableBody.innerHTML += `
             <tr>
-                <td>${element['categoryId']}</td>
-                <td>${element['categoryName']}</td>
-                <td>${str}</td>
-                <td>
-                    <button class="btn event btn-outline-success" data-type="edit" data-bs-toggle="modal" data-bs-target="#staticBackdrop" data-val="${element['categoryId']}">
-                        <i class="fa-solid fa-pencil"></i>  
-                    </button>
-                    <button class="btn event btn-outline-danger" data-type="delete" data-val="${element['categoryId']}">
-                        <i class="fa-solid fa-trash"></i>
-                    </button>
-                </td>
+            <td>${element['categoryId']}</td>
+            <td>${element['categoryName']}</td>
+            <td>${str}</td>
+            <td>
+            <button class="btn event btn-outline-success" data-type="edit" data-bs-toggle="modal" data-bs-target="#staticBackdrop" data-val="${element['categoryId']}">
+            <i class="fa-solid fa-pencil"></i>  
+            </button>
+            <button class="btn event btn-outline-danger" data-type="delete" data-val="${element['categoryId']}">
+            <i class="fa-solid fa-trash"></i>
+            </button>
+            </td>
             </tr>`;
         buttonEventlistner();
     }
+
 }
 
 function addButton(categoryForm, categoryId, categoryName, addDescription) {
+    console.log(data);
     categoryForm.dataset.type = "add";
     categoryId.value = null;
     categoryName.value = null;
@@ -93,6 +102,8 @@ function editCategorySubmitHandler(categoryName, addDescription, categoryId) {
     data[Number(categoryId.value)]['description'] = addDescription.value;
     localStorage.setItem('category', JSON.stringify(data));
 }
+var myModal = new bootstrap.Modal(document.getElementById('staticBackdrop'));
+
 
 categoryForm.addEventListener('submit', (e) => {
     e.preventDefault();
@@ -100,29 +111,33 @@ categoryForm.addEventListener('submit', (e) => {
     const addDescription = document.getElementById('addDescription');
     const categoryId = document.getElementById('categoryId');
     Swal.fire({
-        title: 'Do you want to save the changes?',
+        title: `Do you want to add "${categoryName.value}" category?`,
         showDenyButton: true,
-        showCancelButton: true,
         confirmButtonText: 'Yes',
         denyButtonText: 'No',
         customClass: {
             actions: 'my-actions',
-            cancelButton: 'order-1 right-gap',
             confirmButton: 'order-2',
             denyButton: 'order-3',
         },
     }).then((result) => {
-        if (result.isConfirmed) {
-            (categoryForm.dataset.type == "add") ? addCategorySubmitHandler(categoryName, addDescription)
-                : editCategorySubmitHandler(categoryName, addDescription, categoryId);
-            Swal.fire('Saved!', '', 'success');
-        } else if (result.isDenied) {
-            Swal.fire('Changes are not saved', '', 'info')
+        if (!result.isConfirmed) {
+            Swal.fire(`Category ${categoryName.value} has not been added. `, '', 'info')
+            return;
         }
-    }).then(()=>{
-        displayElements()
-    })
+        if (categoryForm.dataset.type == "add") {
+            addCategorySubmitHandler(categoryName, addDescription);
+        }
+        else {
+            editCategorySubmitHandler(categoryName, addDescription, categoryId);
+        }
+        Swal.fire('Saved!', '', 'success');
+    }).then(() => {
+        resetArr();
+        displayElements(arr)
+        myModal.hide();
 
+    })
 });
 
 function sortAndDisplay(button) {
@@ -146,18 +161,37 @@ function sortAndDisplay(button) {
 }
 
 function deleteButton(button) {
-    for (let i in product) {
-        if (product[i]['category'] == data[button.dataset.val]['categoryId']) {
-            delete product[i];
+    Swal.fire({
+        title: `Do you want to delete "${ data[button.dataset.val]['categoryName']}" category?`,
+        showDenyButton: true,
+        confirmButtonText: 'Yes',
+        denyButtonText: 'No',
+        customClass: {
+            actions: 'my-actions',
+            confirmButton: 'order-2',
+            denyButton: 'order-3',
+        },
+    }).then((result) => {
+        if (!result.isConfirmed) { return; }
+        upadateData();
+        for (let i in product) {
+            if (product[i]['category'] == data[button.dataset.val]['categoryId']) {
+                delete product[i];
+            }
         }
+        delete data[button.dataset.val];
+        jsonString = JSON.stringify(data);
+        localStorage.setItem('category', jsonString);
+        jsonString = JSON.stringify(product);
+        localStorage.setItem('products', jsonString);
+    }).then(() => {
+        resetArr();
+        filterEliments();
+        displayElements(arr)
+        myModal.hide();
 
-    }
-    delete data[button.dataset.val];
-    jsonString = JSON.stringify(data);
-    localStorage.setItem('category', jsonString);
-    jsonString = JSON.stringify(product);
-    localStorage.setItem('products', jsonString);
-    location.reload();
+    })
+
 }
 
 function buttonEventlistner() {
@@ -191,10 +225,15 @@ function buttonEventlistner() {
         })
     });
 }
-filter.addEventListener('input', () => {
-    resetArr();
-    arr = arr.filter(category => category['categoryId'] == Number(filter.value)
+
+function filterEliments() {
+    console.log(data);
+    let filteredArr = arr.filter(category => category['categoryId'] == Number(filter.value)
         || category.categoryName.toLowerCase().includes(String(filter.value.toLowerCase()))
         || category.description.toLowerCase().includes(String(filter.value.toLowerCase())));
-    displayElements(arr);
+    displayElements(filteredArr);
+}
+
+filter.addEventListener('input', () => {
+    filterEliments();
 });
